Update to use hard versions of dependencies
The open ended [1.10.5,) causes maven/ivy to go back to the central repo to look for the latest version every time the build runs.  At least in SBT, this forces the latest version of the SDK to download every time it is released, which has caused problems for our build.  It also makes dependency resolution take an unnecessarily long time.  This update chooses the current latest version of the SDK.