Update PULL_REQUEST_TEMPLATE.md
As part of a PR, it would be good to capture the target of the change. It could be Managed runtime, OCI or both.

RIC implementation hosted in this repository has more features than used in managed runtime (i.e. platform detection).

All changes which are subject to review should have target clearly defined. If a change is targeted to OCI, reviewers should  ensure the change is not affecting managed runtime functionally and performance wise